+++ lib/view/Widgets/widget.dart
@@ -1,0 +1,4 @@
+//barrell file for exporting path 
+
+export 'app_bar.dart';
+export'message_textfield.dart';
