+++ lib/view/screens/screens.dart
@@ -1,0 +1,4 @@
+
+// barell file export path
+export 'chat/chat_screen.dart';
+export 'home/main_screen.dart';
